+++ src/array.cpp
@@ -8,7 +8,7 @@ Tensor ClipByValue(const Tensor& in, float clip_val_min, float clip_val_max)
     assert(clip_val_min <= clip_val_max);
     Tensor out = in;
     
-    for (unsigned int i = 0; i < in.size; ++i) {
+    for (size_t i = 0; i < in.size; ++i) {
         if (in[i] < clip_val_min)
             out[i] = clip_val_min;
         else if (in[i] > clip_val_max)
@@ -22,7 +22,7 @@ static void SetShape(Tensor& in, const std::vector<size_t>& shape)
 {
     in.shape.reserve(shape.size());
 
-    for (unsigned int elem : shape)
+    for (size_t elem : shape)
         assert(elem != 0);
 
     in.shape = std::move(shape);
@@ -31,9 +31,9 @@ static void SetShape(Tensor& in, const std::vector<size_t>& shape)
 static void SetSize(Tensor& in, const std::vector<size_t>& shape)
 {
     if (in.shape.size() > 0) {
-        unsigned int num_elem = 1;
+        size_t num_elem = 1;
 
-        for (unsigned int elem : shape)
+        for (size_t elem : shape)
             num_elem *= elem;
 
         in.size = num_elem;
@@ -53,7 +53,7 @@ static void SetNumChDim(Tensor& in,  const std::vector<size_t>& shape)
     if (in.shape.size() > 0) {
         in.num_ch_dim = 1;
 
-        for (int i = 0; i < shape.size() - 1; ++i)
+        for (size_t i = 0; i < shape.size() - 1; ++i)
             in.num_ch_dim *= shape[i];
 
     } else {
@@ -71,13 +71,13 @@ Tensor Ones(const std::vector<size_t>& shape)
     return out;
 }
 
-Tensor Slice(const Tensor& in, const unsigned int begin, const unsigned int size)
+Tensor Slice(const Tensor& in, const size_t begin, const size_t size)
 {
     assert(begin < in.shape[0] && begin + size <= in.shape[0]);
     Tensor out = Zeros({ size, in.shape.back() });
-    unsigned int idx = 0;
+    size_t idx = 0;
 
-    for (unsigned int i = begin * in.shape.back(); i < (begin * in.shape.back()) + (size * in.shape.back()) ; ++i) {
+    for (size_t i = begin * in.shape.back(); i < (begin * in.shape.back()) + (size * in.shape.back()) ; ++i) {
         out[idx] = in[i];
         ++idx;
     }
